% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_adm.R
\name{aggregate_to_adm}
\alias{aggregate_to_adm}
\title{Aggregate gridded SPAM-C results to administrative unit level}
\usage{
aggregate_to_adm(param, alt_param)
}
\arguments{
\item{param}{Object of type \code{spam_par} that bundles all SPAM parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\value{
SPAM-C results aggregated to the administrative unit level specified in alt_param
}
\description{
\code{aggregate_to_adm} aggregates gridded results of the model specified by \code{alt_param} to the administrative
unit level as specified by the \code{adm_level} parameter in \code{param}. The model described by \code{alt_param} must
be identical to the \code{param} model apart from the \code{adm_level} parameter. By setting this value to a lower
value a less constraint model is created. The output of \code{aggregate_to_adm} can be used as for a comparison
the most detailed subnational information available as a way to validate the model.
}
\examples{
\dontrun{
}
}
