% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_raster.R
\name{align_raster}
\alias{align_raster}
\title{align_raster}
\usage{
align_raster(
  unaligned,
  reference,
  dstfile,
  cutline,
  crop_to_cutline = FALSE,
  r = "bilinear",
  overwrite = TRUE,
  n_threads = "ALL_CPUS"
)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
align_raster
}
