% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cropland_coverage.R
\name{calc_cropland_coverage}
\alias{calc_cropland_coverage}
\title{Calculates the balance between cropland and physical crop area}
\usage{
calc_cropland_coverage(param)
}
\arguments{
\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
\code{calculate_cropland_coverage()} calculates the balance between total
cropland, derived from the synergy cropland mask, and total physical area as
provided by the crop statistics.
}
