% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_statistics.R
\name{check_statistics}
\alias{check_statistics}
\title{Checks the consistency of the subnational statistics}
\usage{
check_statistics(df, param, out = FALSE)
}
\arguments{
\item{df}{data.frame with subnational statistics at various levels in the long format}

\item{out}{logical; should the checking report be returned as output?}
}
\value{
data.frame \code{df} when out is set to \code{TRUE}
}
\description{
Subnational statistics must be of class \code{subnat}
}
