% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adm_map_pdf.R
\name{create_adm_map_pdf}
\alias{create_adm_map_pdf}
\title{Creates a pdf file with maps of administrative unit locations}
\usage{
create_adm_map_pdf(param, font_size = 3)
}
\arguments{
\item{font_size}{Parameter to specify the font size of administrative unit
name that is plotted on the map (default size = 3). If names are long or
administrative unit polygons are small, the labels may get cluttered. Better
results can be obtained by decreasing the font size.}
}
\value{
An rds file with the created grid is saved in the \code{processed_data/maps/grid} folder.
}
\description{
The function creates and saves a pdf file with the location of the
administrative units in the \verb{/processed_data/maps/adm} folder. Depending on the
adm level, maps are produced for level 1 and/or level 2 administrative units.
The function will automatically load the file with the polygon information on
the location of administrative units, that needs to be created first by the
user (see Articles).
}
\examples{
\dontrun{
create_adm_map_pdf(param)
}

}
