% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_tif.R
\name{create_all_tif}
\alias{create_all_tif}
\title{Create tif files for all crop distribution maps produced with \code{mapspamc}}
\usage{
create_all_tif(param)
}
\arguments{
\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
Create tif files for all crop distribution maps produced with \code{mapspamc}
}
