% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Creates \code{mapspamc} folder structure}
\usage{
create_folders(param = NULL)
}
\arguments{
\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
\code{create_mapspamc_folders} creates the folder structure that is needed store raw
data, processed data and parameters for \code{mapspamc}.
}
\details{
\code{create_folders} creates two folders in the \code{mapspamc_path}, set by the user:
mappings and processed_data, and creates a \code{raw_data} folder in the location
as set in \code{mapspamc_par}. In addition, it copies a number of cvs files into the
mappings folder, which contain several data tables that are needed to run the
model and, if needed can be adjusted by the user.
}
\examples{
\dontrun{
create_folders(param)
}

}
