% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_globiom_input.R
\name{create_globiom_input}
\alias{create_globiom_input}
\title{Create land cover and crop distribution gdx files for use with GLOBIOM}
\usage{
create_globiom_input(mapping, lc_map, param)
}
\description{
Create land cover and crop distribution gdx files for use with GLOBIOM
}
