% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_template.R
\name{create_model_template}
\alias{create_model_template}
\title{Creates \code{mapspamc} model template}
\usage{
create_model_template(template_path = NULL)
}
\arguments{
\item{template_path}{folder where template scripts will be copied.If NULL the current working directory will be used.}
}
\description{
\code{create_model_template} creates a folder structure with template R scripts to implement the various
steps to produce crop distribution maps with \code{mapspamc}.
}
\details{
\verb{create_model template()} creates nine folders in the \code{template_path}, set by the user. The folders
contain template R scripts to implement the six steps to generate crop distribution
maps with \code{mapspamc}, including: (1) model setup, (2) pre-processing, (3) model preparation,
(4) running the model, (5) post-processing and (6) model validation. The user
can use the template as a starting point for a country case-study.

To run the template scripts, they have to be copied into an RStudio project. We recommend to
create such a project first and then use \code{create_model_template()} to add the scripts to
the RStudio project. Note that \code{create_model_template()} only copies files if they do not exist
yet. In this way, the user cannot accidently replace already modified scripts.
}
\examples{
\dontrun{
create_model_template("c:/temp")
}

}
