% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_inputs.R
\name{harmonize_inputs}
\alias{harmonize_inputs}
\title{Harmonizes crop statistics, cropland and irrigated area extent}
\usage{
harmonize_inputs(
  param,
  cl_slackp = 0.05,
  cl_slackn = 5,
  ia_slackp = 0.05,
  ia_slackn = 5
)
}
\arguments{
\item{cl_slackp}{percentage of total cropland that will be added when comparing
physical area statistics to cropland extent.}

\item{cl_slackn}{times the area of the largest grid cell will be added when comparing
physical area statistics to cropland extent.}

\item{ia_slackp}{percentage of total irrigated area that will be added when comparing
irrigated area statistics to irrigated area extent.}

\item{ia_slackn}{times the area of the largest grid cell will be added when comparing
physical area statistics to irrigated area extent.}
}
\description{
Function that compares and harmonizes the subnational physical area with gridded information on the
location and extend of cropland and irrigated area.
}
\examples{
harmonize_inputs(param, cl_slackp = 0.05, cl_slackn = 5, ia_slackp = 0.05, ia_slackn = 5)

}
