% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapspamc.R
\docType{package}
\name{mapspamc-package}
\alias{mapspamc}
\alias{mapspamc-package}
\title{mapspamc: An R package to create crop distribution maps}
\description{
The aim of the mapspamc R package is to facilitate the creation of country level crop distribution maps. mapspamc provides the necessary infrastructure to run the Spatial Production Allocation Model for Country level studies (SPAMc). The model builds on the global version of SPAM (You et al., 2006; You et al., 2009 et al.; You et al., 2014 and Yu et al., 2020), which use a cross-entropy optimization approach to 'pixelate' national and subnational crop statistics on a spatial grid at a resolution of 5 arcmin. SPAMc was specifically developed to support country level analysis and makes it possible to incorporate national sources of information and create maps at a higher resolution of 30 arcsec.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michielvandijk/mapspamc}
  \item Report bugs at \url{https://github.com/michielvandijk/mapspamc/issues}
}

}
\author{
\strong{Maintainer}: Michiel Van Dijk \email{michiel.vandijk@wur.nl} (\href{https://orcid.org/0000-0002-5207-7304}{ORCID})

}
\keyword{internal}
