% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Sum of vector elements but now NA + NA = NA not 0 as in sum}
\usage{
plus(x, na.rm = F)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The sum of \code{x}
}
\description{
\code{plus} Returns the sum of all values provided as arguments but ensures \code{NA} +
\code{NA} = \code{NA}.
}
\details{
This function is the same as \code{sum}() but if \code{na.rm} is \code{FALSE} and all input
values are \code{NA}, it will return \code{NA} instead of 0.
}
\examples{
plus(1:10)
plus(c(NA, NA))
plus(c(NA, NA), na.rm = T)
}
