% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cropland_tp1.R
\name{prepare_cropland_tp1}
\alias{prepare_cropland_tp1}
\title{Prepares the synergy cropland area for SPAM including tp1}
\usage{
prepare_cropland_tp1(param)
}
\arguments{
\item{param}{Object of type \code{spam_par} that bundles all SPAM parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
Prepares the synergy cropland area for SPAM including tp1
}
\examples{

}
