% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_irrigated_area.R
\name{prepare_irrigated_area}
\alias{prepare_irrigated_area}
\title{Prepares synergy irrigated area extent}
\usage{
prepare_irrigated_area(param)
}
\arguments{
\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
Combines all elements of the synergy irrigated area extent (maximum irrigated area and rank),
with information on the location of the subnational units and fixes potential inconsistencies
(e.g. irrigated area larger than grid cell size).
}
\examples{
\dontrun{
prepare_irrigated_area(param)
}

}
