% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_adm_map.R
\name{rasterize_adm_map}
\alias{rasterize_adm_map}
\title{Rasterizes the map with the location of the subnational administrative units}
\usage{
rasterize_adm_map(param)
}
\value{
RasterLayer
}
\description{
For several internal operations, \code{mapspamc} needs a rasterized version of the map
with locations of the subnational administrative units. This function creates
this map and saves it into the \verb{/processed_data/maps/adm/} folder.
}
\examples{
\dontrun{
rasterize_adm_map(param)
}
}
