% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaggregate_statistics.R
\name{reaggregate_statistics}
\alias{reaggregate_statistics}
\title{Reaggregates subnational statistics from the bottom up so they are consistent}
\usage{
reaggregate_statistics(df, param)
}
\arguments{
\item{df}{tbl or data.frame}
}
\value{
same class as \code{df}.
}
\description{
Reaggregates subnational statistics from the bottom up so they are consistent
}
