% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_calc_cropland_coverage.R
\name{split_calc_cropland_coverage}
\alias{split_calc_cropland_coverage}
\title{Calculates cropland coverage in line with solve level}
\usage{
split_calc_cropland_coverage(ac, param)
}
\description{
Calculates cropland coverage in line with solve level
}
