% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_results.R
\name{view_results}
\alias{view_results}
\title{Compares crop distribution maps using a panel for each production system}
\usage{
view_results(crp, var, param, viewer = TRUE, polygon = TRUE)
}
\arguments{
\item{crp}{Character. Crop for which the maps are shows. \code{crp}  has to be
one of the \code{mapspamc}  four letter crop codes.}

\item{var}{Character. The variable to be plotted. \code{var} has to be physical
area (\code{"pa"}) or harvested area (\code{"ha"}).}

\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}

\item{viewer}{Logical. Should the default web browers be used to show the
maps? \code{FALSE} will show the maps in the RStudio viewer pane.}

\item{polygon}{Logical; should the country polygon be overlayed?}
}
\description{
To quickly inspect the \code{mapspamc} results, \code{view_results()} shows crop
distribution maps for a selected crop using a panel for each production system.
This function works after \code{combine_results()} is run. There is no need to
run \code{create_all_tif()}. The maps are visualized using
\href{https://leafletjs.com/}{leafletjs}, which makes it possible to select a
number of background tiles (e.g. OpenStreetMap).
}
\examples{
\dontrun{
view_results(crop = "maiz", var = "ha", viewer = FALSE, polygon = FALSE)
}

}
