% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_adm.R
\name{aggregate_to_adm}
\alias{aggregate_to_adm}
\title{Aggregates gridded results of alternative model to the administrative unit level of the target model}
\usage{
aggregate_to_adm(param, alt_param)
}
\arguments{
\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}

\item{alt_param}{Object of type spamc_par that bundles all \code{mapspamc} parameters, including core model folders,
alpha-3 country code, year, spatial resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of model.}
}
\value{
data.frame with \code{mapspamc} results aggregated to the administrative unit level specified in
\code{alt_param}.
}
\description{
\code{aggregate_to_adm} aggregates gridded results of an alternative model specified by \code{alt_param}
to the administrative unit level determined by \code{adm_level} of the target model specified by \code{param}.
The model described by \code{alt_param} must be identical to the target model apart from the \code{adm_level}
parameter, which is set to a less-detailed level. A comparison of the output of \code{aggregate_to_adm} with
subnational information at the most detailed level available is a validation of the alternative model and,
indirectly, also of the target model.
}
\examples{
\dontrun{
aggregate_to_adm(param, alt_param)
}

}
