% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{Combines GAMS output into one R rds file}
\usage{
combine_results(param, cut = 1e-04, out = FALSE)
}
\arguments{
\item{cut}{numeric. Sets allocation smaller than a certain value to 0. The default is 0.0001 (1 m2).}

\item{out}{logical; should the results be returned as output?}
}
\description{
Combines the GAMs results that are saved in one (\code{solve_level = 0})  or multiple (\code{solve_level = 1}) gdx files into one single rds file, that can be
easily loaded into R. The file is saved in the \code{processed_data/results} folder.
}
\examples{
\dontrun{
combine_results(param)
}#'
}
