% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_tif.R
\name{create_all_tif}
\alias{create_all_tif}
\title{Creates tif files for all crop distribution maps produced by the model}
\usage{
create_all_tif(param)
}
\description{
The function creates tif files with maps for all crop and system combinations for which area information
was allocated by the model. Maps are both created for harvested (ha) and physical (pa) area. Files are
saved in the \verb{/processed_data/results/\{model\}/maps} folder.
}
