% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistics_template.R
\name{create_statistics_template}
\alias{create_statistics_template}
\title{Creates template for subnational crop statistics}
\usage{
create_statistics_template(type, param)
}
\arguments{
\item{type}{Character vector that refers to the type of template that needs to
be created. See details for allowed input.}
}
\description{
To support the preparation of the subnational statistics,
\code{create_statistics_template()} can create three types of data templates:
\itemize{
\item ha for harvest area statistics
\item fs for farming system share
\item ci for cropping intensity.
}
}
\details{
The function requires information on how the different administrative unit
levels are nested. This file needs to be created first by running
\code{create_adm_list()}.

The dimensions of the templates are determined by the parameters set by means of
\code{mapspamc_par()}. The ha template dimension is determined by adm level parameter,
while the dimensions of the fs and cs templates (which are identical) are determined
by the solve level parameter.
}
\examples{
create_statistics_template(type = "ha", param)

}
