% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistics_template.R
\name{create_statistics_template}
\alias{create_statistics_template}
\title{Create template for raw subnational crop statistics}
\usage{
create_statistics_template(type, param)
}
\arguments{
\item{type}{Character vector that refers to the type of template that needs to
be created. See details for allowed input.}

\item{param}{Object of type \code{spam_par} that bundles all SPAM parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
To support the preparation of the subnational statistics,
\code{create_raw_statistics_template()} can create three types of data templates:
\itemize{
\item ha for harvest area statistics
\item fs for farming system share
\item ci for cropping intensity.
}
}
\details{
The function requires information on how the different administrative unit
levels are nested. This file needs to be created first by running
\code{create_adm_list()}.

The dimensions of the ha template are determined by adm level set by
\code{spam_par()}, while the dimensions of the fs and cs templates (which are
identical) are determined by the solve level parameter.
}
\examples{
create_statistics_template(type = "ha", param)

}
