% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_inputs.R
\name{harmonize_inputs}
\alias{harmonize_inputs}
\title{Harmonizes the subnational statistics, the cropland extent and the irrigated area}
\usage{
harmonize_inputs(param, cl_slackp = 0.05, cl_slackn = 5, ia_slackp = 0.05)
}
\arguments{
\item{param}{Object of type \code{mapspamc_par} that bundles all \code{mapspamc} parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
Harmonizes the subnational statistics, the cropland extent and the irrigated area
}
\examples{

}
