% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cropland.R
\name{prepare_cropland}
\alias{prepare_cropland}
\title{Prepares synergy cropland extent}
\usage{
prepare_cropland(param)
}
\description{
Combines all elements of the synergy cropland extent (median and maximum cropland per grid cell and rank),
with information on the location of the subnational units and fixes potential inconsistencies
(e.g. cropland area larger than grid cell size).
}
\examples{

}
