% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_irrigated_area.R
\name{prepare_irrigated_area}
\alias{prepare_irrigated_area}
\title{Prepares synergy irrigated area extent}
\usage{
prepare_irrigated_area(param)
}
\description{
Combines all elements of the synergy irrigated area extent (maximum irrigated area and rank),
with information on the location of the subnational units and fixes potential inconsistencies
(e.g. irrigated area larger than grid cell size).
}
