% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_irrigated_area.R
\name{prepare_irrigated_area}
\alias{prepare_irrigated_area}
\title{Prepares the synergy irrigated area for SPAM}
\usage{
prepare_irrigated_area(param)
}
\arguments{
\item{param}{Object of type \code{spam_par} that bundles all SPAM parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\description{
Prepares the synergy irrigated area for SPAM
}
\examples{

}
