% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_priors_and_scores.R
\name{prepare_priors_and_scores}
\alias{prepare_priors_and_scores}
\title{Prepares priors and scores for all farming systems, crops and grid cells}
\usage{
prepare_priors_and_scores(param)
}
\description{
Prepares priors and scores for all farming systems, crops and grid cells
}
