% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_adm_map.R
\name{rasterize_adm_map}
\alias{rasterize_adm_map}
\title{Rasterizes the map with the location of the subnational administrative units}
\usage{
rasterize_adm_map(param)
}
\arguments{
\item{param}{Object of type \code{spam_par} that bundles all SPAM parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved and type of
model.}
}
\value{
RasterLayer
}
\description{
For several internal operations, SPAMc needs a rasterized version of the map
with locations of the subnational administrative units. This function creates
this map and saves it into the \verb{/processed_data/maps/adm/} folder.
}
\examples{
\dontrun{
rasterize_adm_map(param)
}
}
